requirements.txt
Django==1.11.8
pytz==2017.3
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
